# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Rental Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    get:
      operationId: getQuotes
      tags: [quotes]
      summary: Get all custom quotes
      responses:
        "200":
          description: List of quotes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Create a custom quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuote"
      responses:
        "201":
          description: Created quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
  /quotes/{id}:
    put:
      operationId: updateQuote
      tags: [quotes]
      summary: Update a quote
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateQuote"
      responses:
        "200":
          description: Updated quote
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
...
components:
  schemas:
...
    Quote:
      type: object
      properties:
        id:
          type: number
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/QuoteItem"
        startDate:
          type: string
        endDate:
          type: string
        subtotal:
          type: number
        discount:
          type: number
        totalPrice:
          type: number
        status:
          type: string
          enum: [draft, sent, accepted, declined, expired]
        notes:
          type: string
          nullable: true
        validUntil:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - customerName
        - customerEmail
        - items
        - startDate
        - endDate
        - subtotal
        - discount
        - totalPrice
        - status
        - createdAt
        - updatedAt
    QuoteItem:
      type: object
      properties:
        listingId:
          type: number
        listingTitle:
          type: string
        quantity:
          type: number
        pricePerDay:
          type: number
        days:
          type: number
        subtotal:
          type: number
      required:
        - listingId
        - listingTitle
        - quantity
        - pricePerDay
        - days
        - subtotal
    CreateQuote:
      type: object
      properties:
        customerName:
          type: string
        customerEmail:
          type: string
        customerPhone:
          type: string
          nullable: true
        items:
          type: array
          items:
            type: object
            properties:
              listingId:
                type: number
              quantity:
                type: number
              pricePerDay:
                type: number
              days:
                type: number
            required: [listingId, quantity, pricePerDay, days]
        startDate:
          type: string
        endDate:
          type: string
        discount:
          type: number
        notes:
          type: string
          nullable: true
        validUntil:
          type: string
          nullable: true
      required:
        - customerName
        - customerEmail
        - items
        - startDate
        - endDate
        - discount
    UpdateQuote:
      type: object
      properties:
        status:
          type: string
          enum: [draft, sent, accepted, declined, expired]
        discount:
          type: number
        notes:
          type: string
          nullable: true
        validUntil:
          type: string
          nullable: true
...